Check for active article flags not just flags
