+++ packages/provider-test-tools/fixtures/article.json
@@ -1366,7 +1366,7 @@
   "expirableFlags": [
     {
       "type": "BREAKING",
-      "expiryTime": "2030-03-13T13:00:00.000Z"
+      "expiryTime": "2020-03-13T13:00:00.000Z"
     },
     {
       "type": "NEW",
@@ -1386,7 +1386,7 @@
     },
     {
       "type": "LIVE",
-      "expiryTime": "2030-03-13T13:00:00.000Z"
+      "expiryTime": "2020-03-13T13:00:00.000Z"
     }
   ],
   "hasVideo": true,
